+++ service/engine/storage/storage.h
@@ -14,8 +14,8 @@ class Storage {
  public:
   virtual ~Storage() {};
   virtual StoragePage *AllocatePage() = 0;
-  virtual StoragePage *GetPage(uint64_t index) = 0;
-  virtual uint64_t GetPagesCount() = 0;
+  virtual StoragePage *GetPage(int index) = 0;
+  virtual int GetPagesCount() = 0;
   virtual void Flush() = 0;
 };
